requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1
pdfminer.six
spacy
pyyaml<=5.1
wget
# conda install -c anaconda beautifulsoup4
lxml